Allow to preserve certain environment variables
The idea is to list variables, which should be preserved from a parent
environment and pass them as for prebuild*.sh scripts as well as to the
rpmbuild / debuild environment.

There were two attempts to solve the problem: `VAR_` prefix (#95,
pushed) and `PACKPASS_` prefix (#101, not pushed). Why I decided to
implement another one? We have a testing tool, which is tuneable using
environment variables. We have different jobs in CI, which invokes the
tool. Some CI jobs use this tool directly, but some run packpack, which
calls it under rpmbuild or debuild. My wish it to set an environment
variable once (in CI settings) and use it everywhere: whether packpack
is used or not. I don't want to hack CI scripts or specs to pass the
variables to the tool.

rpmbuild inherits a parent environment (maybe except certain variables),
so it does not require any changes aside passing a variable to the
docker's environment.

debuild runs build scripts in a clean environment (that is nice) and so
requires to explicitly list all variables to preserve.

We can either set DEBUILD_PRESERVE_ENVVARS configuration variable
(`~/.devscripts`) or pass a bunch of --preserve-envvar command line
options. I choose the second way in order to show the variables in the
build log.

The problem comes from [1].

[1]: https://github.com/tarantool/test-run/issues/258

Fixes #100
Supersedes #101